Extract make_proxy_cm and isinstance_class into shared tests/helpers.py
Both were byte-for-byte duplicated across several test files
(make_proxy_cm in 5, isinstance_class in 2). Centralizing them keeps
the async-context-manager and fake-isinstance-class construction
consistent instead of independently maintained copies.
