+++ tests/robotic/scripts/test_autofocus.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import itertools
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
@@ -8,19 +7,7 @@ import pytest
 from pyobs.robotic.scheduler.targets import SiderealTarget
 from pyobs.robotic.scripts.imaging.autofocus import AutoFocusScript
 from pyobs.robotic.task import TaskData
-
-_mock_class_counter = itertools.count()
-
-
-def _isinstance_class(name: str, interfaces: list[type]) -> type:
-    """Build a fresh class purely for isinstance() checks against a MagicMock.
-
-    Uniquely named per call: an unqualified name built only from interface
-    bases would otherwise land in the Interface registry (it's structurally
-    indistinguishable from a real composite interface) and collide with
-    itself across repeated calls, since each call creates a new class object.
-    """
-    return type(f"_Mock{name}{next(_mock_class_counter)}", tuple(interfaces), {})
+from tests.helpers import isinstance_class, make_proxy_cm
 
 
 def make_script(**kwargs) -> AutoFocusScript:
@@ -45,7 +32,7 @@ def make_telescope(ready=True, is_motion=True) -> MagicMock:
     tel.get_state = MagicMock(return_value=ReadyState(ready=ready))
     tel.move_radec = AsyncMock()
     tel.stop_motion = AsyncMock()
-    tel.__class__ = _isinstance_class("Telescope", interfaces)
+    tel.__class__ = isinstance_class("Telescope", interfaces)
     return tel
 
 
@@ -55,13 +42,6 @@ def make_autofocus() -> MagicMock:
     af = MagicMock(spec=[IAutoFocus])
     af.auto_focus = AsyncMock()
     return af
-
-
-def make_proxy_cm(value: object) -> MagicMock:
-    cm = MagicMock()
-    cm.__aenter__ = AsyncMock(return_value=value)
-    cm.__aexit__ = AsyncMock(return_value=None)
-    return cm
 
 
 # ── can_run ───────────────────────────────────────────────────────────────────
