+++ tests/robotic/scripts/test_darkbias.py
@@ -1,25 +1,12 @@
 from __future__ import annotations
 
-import itertools
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
 from pyobs.robotic.scripts.calibration.darkbias import DarkBiasScript
 from pyobs.utils.enums import ImageType
-
-_mock_class_counter = itertools.count()
-
-
-def _isinstance_class(name: str, interfaces: list[type]) -> type:
-    """Build a fresh class purely for isinstance() checks against a MagicMock.
-
-    Uniquely named per call: an unqualified name built only from interface
-    bases would otherwise land in the Interface registry (it's structurally
-    indistinguishable from a real composite interface) and collide with
-    itself across repeated calls, since each call creates a new class object.
-    """
-    return type(f"_Mock{name}{next(_mock_class_counter)}", tuple(interfaces), {})
+from tests.helpers import isinstance_class, make_proxy_cm
 
 
 def make_script(**kwargs) -> DarkBiasScript:
@@ -55,15 +42,8 @@ def make_camera(
     camera.grab_data = AsyncMock()
 
     # make isinstance checks work
-    camera.__class__ = _isinstance_class("Camera", interfaces)
+    camera.__class__ = isinstance_class("Camera", interfaces)
     return camera
-
-
-def make_proxy_cm(value: object) -> MagicMock:
-    cm = MagicMock()
-    cm.__aenter__ = AsyncMock(return_value=value)
-    cm.__aexit__ = AsyncMock(return_value=None)
-    return cm
 
 
 def setup_run_comm(script: DarkBiasScript, camera: MagicMock, binning_cam=..., window_cam=...) -> None:
